+++ spec/spec_helper.rb
@@ -7,8 +7,6 @@ $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 require 'rspec'
 require 'openstudio-analysis'
 
-
-
 RSpec.configure do |config|
   # Use color in STDOUT
   config.color = true